Added Ableton Push 2
Hi Sema.
While adding this, I cleaned up the list a bit for better oversight.
Removed a space right after the f7 of Roland XV-3080.
Akai Push 1 gives a reply too, but it contains a unique device serial. I changed the serial to 12345-12345678 for this example.   
f0 7e 00 06 02 47 15 00 19 00 01 01 06 00 00 00 00 00 31 32 33 34 35 2d 31 32 33 34 35 36 37 38 00 00 f7

Thanks for sharing this project!

- Bjorn